desktop/requirements.txt
Pillow>=10.0.0
python-docx>=1.1.0
